Rename util prompt module -> input
